deleteGSE {interGEO}	R Documentation

deleteGSE

Description

Deletes a series matrix with the "DELETE SM" option. It launches deleteGSE function. If no series matrix is either selected or available, an error message is launched, but the program keeps running!

Usage

deleteGSE(env0)

Details

Eventually, it uploads global environment.

[Package interGEO version 0.1 Index]