getData {interGEO}	R Documentation

getData

Description

Creates a new expression-data matrix associated to selected dataset with the "CREATE DATA" option. If no series matrix is either selected or available, an error message is launched, but the program keeps running!

Usage

getData(env0)

Details

Eventually, it uploads the expression-data matrix (class: ExpressionSet) to global environment, and checks if there are some other associated series matrices.

[Package interGEO version 0.1 Index]