getFactor {interGEO}	R Documentation

getFactor

Description

Adds a factor row to a expression-data matrix

Usage

getFactor(env0, env.main)

Details

Factor is added according to informations contained in pData(series.matrix). You can rename both factor and levels in factor, add multiple factors at once and merge factors with a factorial product ("MERGE" option). You can't add a factor whose name is already present as a rowname in the data matrix. In this case, the program informs you with a warning message. Eventually, it uploads data object to global environment.

[Package interGEO version 0.1 Index]