getGSE {interGEO}	R Documentation

getGSE

Description

Imports series matrices with the "LOAD DATASET" option in iGEO. If the name does not correspond to any dataset in GEO, an error message is launched, but the program keeps running! If a GSE file contains many series matrices, all of them are imported. You can rename each series matrix.

Usage

getGSE(env0)

Details

Eventually, it uploads series matrices (class: ExpressionSet) to global environment, and checks if there are already some associated data.

[Package interGEO version 0.1 Index]