iGEO {interGEO}	R Documentation

iGEO

Description

Main function. Launches the GUI and all other functions are called by this one.

Usage

iGEO()

Details

The left panel shows all the GEO "series matrices" available in workspace. By selecting one of those series matrices, a list of all the related expression-data matrices appears in the right panel. Association series matrix-data is performed according to column name matches. The dimension of the data matrix (which can be different from the raw one obtained from exprs(series.matrix)) is shown. Data objects can even contain factors in some rows. It will be an important feature in the "WORKONDATA" option. Each time you select an object, such object moves at the top of its panel.

Import series matrices with the "LOAD DATASET" option. It launches getGSE function. If the name does not correspond to any dataset in GEO, an error message is launched, but the program keeps running! If a GSE file contains many series matrices, all of them are imported. You can rename each series matrix.

Create a new expression-data matrix associated to selected dataset with the "CREATE DATA" option. It launches getData function. If no series matrix is either selected or available, an error message is launched, but the program keeps running!

Delete a series matrix with the "DELETE SM" option. It launches deleteGSE function. If no series matrix is either selected or available, an error message is launched, but the program keeps running!

Perform basic preliminary operations on data with the "WORK ON DATA" option. It launches workonData function. If no data is either selected or available, an error message is launched, but the program keeps running!

Scale data column with with the "WORK ON DATA" option. It launches scaleData function. It is not a simple scale function from the Base library, because of the fact that data matrices can also contain factors in some rows. If no data is either selected or available, an error message is launched, but the program keeps running!

Delete an expression-data matrix with the "DELETE DATA" option. It launches deleteData function. If no data is either selected or available, an error message is launched, but the program keeps running!

[Package interGEO version 0.1 Index]